+++ lib/dartapi.dart
@@ -1,6 +1,7 @@
 export 'cli/create_project.dart';
 export 'cli/run_server.dart';
 export 'cli/generate_controllers.dart';
+export 'cli/generate_resource.dart';
 export 'cli/generate_docs.dart';
 export 'cli/generate_migration.dart';
 export 'cli/run_migrations.dart';